modify:productStatusEnum Updated and Config Product Seeder to Seed 500 Row

+++ database/migrations/2026_02_25_125807_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description')->nullable();
             $table->decimal('price', 8, 2);
-            $table->decimal('stock_quantity', 8, 2);
+            $table->integer('stock_quantity');
             $table->integer('low_stock_threshold')->default(10);
             $table->enum('status', ['active', 'inactive', 'discontinued'])->default('active');
             $table->softDeletes();
